Fix move tab file length budget
