+++ .github/swift-file-length-budget.tsv
@@ -3,9 +3,9 @@
 # Reduce counts as files shrink. CI fails if tracked files exceed this budget.
 19963	CLI/cmux.swift
 16745	Sources/TerminalController.swift
-15874	Sources/ContentView.swift
+15889	Sources/ContentView.swift
 14157	Sources/AppDelegate.swift
-13686	Sources/Workspace.swift
+13694	Sources/Workspace.swift
 13354	Sources/GhosttyTerminalView.swift
 10603	Sources/Panels/BrowserPanel.swift
 8406	Sources/cmuxApp.swift
